+++ CHANGELOG.md
@@ -3,5 +3,7 @@ Thank you!
 Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED
 
 -----------------
+- Desert Village changes are now toggled with a config.
+  - This adds a structure processor to convert the Jungle wood set into the Palm wood set.
 - Reorganized data generation for configured and placed features in caves into their own classes.
 - Changed the `wilderwild:blank_shut_up` feature into the `wilderwild:empty` feature, which now generates nothing instead of a single Water block.